An OSM ID has to be a bigint

+++ data/osmstats.sql
@@ -43,7 +43,7 @@ CREATE TYPE objtype AS ENUM ('node', 'way', 'relation');
 CREATE TYPE status AS ENUM ('notags', 'complete', 'incomplete', 'badvalue', 'correct', 'badgeom');
 
 CREATE TABLE public.validation (
-    osm_id integer,
+    osm_id bigint,
     type objtype,
     status status[],
     timestamp timestamp with time zone,
